+++ CHANGELOG.rst
@@ -1,6 +1,15 @@
 Changelog
 *********
 
+0.6.0 (2021-05-16)
+------------------
+
+* Update Read the Docs.
+* Add ``fuc.api.pyvcf.VcfFrame.markmiss_ad`` method.
+* Add ``full`` argument to ``fuc.api.pyvcf.VcfFrame.markmiss_ad/af/dp`` methods.
+* Add ``fucfind`` command.
+* Update ``dfsum`` command.
+
 0.5.0 (2021-05-06)
 ------------------